+++ app/views/users/filters/_by_follow_up_statuses_button.html.erb
@@ -1,7 +1,7 @@
 <div class="mb-1 me-2 position-relative text-center d-flex align-items-center justify-content-center"
       data-controller="users-list-filter dropdown--show-options"
       data-users-list-filter-param-name-value="follow_up_statuses">
-  <% selected_values_count = params[:follow_up_statuses]&.size || 0 %>
+  <% selected_values_count = selected_follow_up_statuses_count(@statuses_count) %>
   <button type="button"
           class="btn btn-grey dropdown-toggle <%= "border-blue" if selected_values_count > 0 %>"
           data-action="dropdown--show-options#toggle"
