+++ app/views/users/users_list_indicators/_follow_up_status_filter_shortcut.html.erb
@@ -1,4 +1,4 @@
-<% active = follow_up_statuses_shortcut_active?(statuses) %>
+<% active = follow_up_statuses_filter_active?(statuses) %>
 <%= link_to structure_users_path(url_params.except(:page).merge(follow_up_statuses: statuses)),
             class: "status-shortcut status-shortcut--#{color} rounded bg-white p-2 " \
                    "d-inline-flex align-items-center gap-2 text-decoration-none " \
